src/backend/requirements.txt
Django
psycopg2
psycopg2-binary
gunicorn==19.6.0
django-oauth-toolkit
django-cors-middleware
djangorestframework
djangorestframework-jwt
cryptography
coverage
python-jose
paramiko
idna==2.7
requests
